seperate block-dispatcher and worker-dispatcher, added schema reloads
