+++ packages/node/package.json
@@ -30,7 +30,7 @@
     "@polkadot/api": "9.5.2",
     "@subql/common": "latest",
     "@subql/common-cosmos": "workspace:*",
-    "@subql/node-core": "1.3.3",
+    "@subql/node-core": "1.4.1",
     "@subql/types-cosmos": "workspace:*",
     "@subql/utils": "latest",
     "@subql/x-merkle-mountain-range": "2.0.0-0.1.2",
