+++ packages/node/src/indexer/fetch.service.ts
@@ -40,10 +40,10 @@ import { SubqlProjectDs, SubqueryProject } from '../configure/SubqueryProject';
 import * as CosmosUtil from '../utils/cosmos';
 import { isBaseHandler, isCustomHandler } from '../utils/project';
 import { ApiService, CosmosClient } from './api.service';
+import { IBlockDispatcher } from './blockDispatcher/block-dispatcher.service';
 import { DictionaryService } from './dictionary.service';
 import { DsProcessorService } from './ds-processor.service';
 import { DynamicDsService } from './dynamic-ds.service';
-import { IBlockDispatcher } from './worker/block-dispatcher.service';
 
 const logger = getLogger('fetch');
 let BLOCK_TIME_VARIANCE = 5000; //ms
@@ -117,10 +117,9 @@ export class FetchService implements OnApplicationShutdown {
   private latestBestHeight: number;
   private latestFinalizedHeight: number;
   private isShutdown = false;
-  private useDictionary: boolean;
-  private dictionaryQueryEntries?: DictionaryQueryEntry[];
   private batchSizeScale: number;
   private templateDynamicDatasouces: SubqlProjectDs[];
+  private dictionaryGenesisMatches = true;
 
   constructor(
     private apiService: ApiService,
@@ -155,13 +154,20 @@ export class FetchService implements OnApplicationShutdown {
       await this.dynamicDsService.getDynamicDatasources();
   }
 
-  getDictionaryQueryEntries(): DictionaryQueryEntry[] {
+  buildDictionaryQueryEntries(startBlock: number): DictionaryQueryEntry[] {
     const queryEntries: DictionaryQueryEntry[] = [];
-
     const dataSources = this.project.dataSources.filter((ds) =>
       isRuntimeDataSourceV0_3_0(ds),
     );
-    for (const ds of dataSources.concat(this.templateDynamicDatasouces)) {
+
+    // Only run the ds that is equal or less than startBlock
+    // sort array from lowest ds.startBlock to highest
+    const filteredDs = dataSources
+      .concat(this.templateDynamicDatasouces)
+      .filter((ds) => ds.startBlock <= startBlock)
+      .sort((a, b) => a.startBlock - b.startBlock);
+
+    for (const ds of filteredDs) {
       const plugin = isCustomCosmosDs(ds)
         ? this.dsProcessorService.getDsProcessor(ds)
         : undefined;
@@ -190,7 +196,8 @@ export class FetchService implements OnApplicationShutdown {
               .filter,
           ];
         }
-        filterList = filterList.filter((f) => f);
+        // Filter out any undefined
+        filterList = filterList.filter(Boolean);
         if (!filterList.length) return [];
         switch (baseHandlerKind) {
           case SubqlCosmosHandlerKind.Block:
@@ -235,10 +242,20 @@ export class FetchService implements OnApplicationShutdown {
   }
 
   updateDictionary(): void {
-    this.dictionaryQueryEntries = this.getDictionaryQueryEntries();
-    this.useDictionary =
-      !!this.dictionaryQueryEntries?.length &&
-      !!this.project.network.dictionary;
+    this.dictionaryService.buildDictionaryEntryMap<SubqlProjectDs>(
+      this.project.dataSources.concat(this.templateDynamicDatasouces),
+      this.buildDictionaryQueryEntries.bind(this),
+    );
+  }
+  private get useDictionary(): boolean {
+    return (
+      !!this.project.network.dictionary &&
+      this.dictionaryGenesisMatches &&
+      !!this.dictionaryService.getDictionaryQueryEntries(
+        this.blockDispatcher.latestBufferedHeight ??
+          Math.min(...this.project.dataSources.map((ds) => ds.startBlock)),
+      ).length
+    );
   }
 
   async init(startHeight: number): Promise<void> {
@@ -363,12 +380,12 @@ export class FetchService implements OnApplicationShutdown {
           queryEndBlock,
         );
         try {
-          const dictionary = await this.dictionaryService.getDictionary(
-            startBlockHeight,
-            queryEndBlock,
-            scaledBatchSize,
-            this.dictionaryQueryEntries,
-          );
+          const dictionary =
+            await this.dictionaryService.scopedDictionaryEntries(
+              startBlockHeight,
+              queryEndBlock,
+              scaledBatchSize,
+            );
 
           if (startBlockHeight !== getStartBlockHeight()) {
             logger.debug(
@@ -447,7 +464,7 @@ export class FetchService implements OnApplicationShutdown {
         logger.error(
           'The dictionary that you have specified does not match the chain you are indexing, it will be ignored. Please update your project manifest to reference the correct dictionary',
         );
-        this.useDictionary = false;
+        this.dictionaryGenesisMatches = false;
         this.eventEmitter.emit(IndexerEvent.UsingDictionary, {
           value: Number(this.useDictionary),
         });
